rename files with new action name
