+++ actions/cuda_snap/sw/action_create_vector.c
